# openapi/atuin-openapi.yaml
openapi: "3.0.0"
info:
  version: 0.1.0
  title: Natuin API
paths:
...
  /sync/count:
    get:
      summary: get sync count
      x-eov-operation-handler: SyncController
      operationId: getCount
      tags:
        - sync
      security:
        - tokenAuth: [ ]
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: object
                required:
                  - count
                properties:
                  count:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
...
  /sync/history:
    get:
      summary: Get sync history
      x-eov-operation-handler: SyncController
      operationId: getHistory
      tags:
        - sync
      security:
        - tokenAuth: [ ]
      parameters:
        - name: sync_ts
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: last sync timestamp
        - name: history_ts
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: history timestamp
        - name: host
          in: query
          allowEmptyValue: true
          schema:
            type: string
            nullable: true
          description: hostname to exclude
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: object
                required:
                  - history
                properties:
                  history:
                    type: array
                    items:
                      type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /sync/calendar/{focus}:
    get:
      summary: Get sync calendar
      x-eov-operation-handler: SyncController
      operationId: calendar
      tags:
        - sync
      security:
        - tokenAuth: [ ]
      parameters:
        - name: focus
          in: path
          required: true
          schema:
            type: string
            enum:
              - year
              - month
              - day
          description: The focus of the calendar
        - name: year
          in: query
          required: false
          schema:
            type: integer
            minimum: 2020
            maximum: 2100
          description: The year
        - name: month
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: The month
        - name: tz
          in: query
          required: false
          schema:
            type: string
            pattern: '^[A-Za-z0-9\/\+\-]+$'
          description: The timezone
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
...
components:
  securitySchemes:
    tokenAuth:
      type: http
      scheme: Token
  schemas:
...
    BasicErrorModel:
      type: object
      required:
        - status
        - reason
      properties:
        status:
          type: integer
          format: int32
        reason:
          type: string
...
    UnauthorizedError:
      allOf:
        - $ref: '#/components/schemas/BasicErrorModel'
        - type: object
          properties:
            status:
              example: 401
            reason:
              example: Unauthorized
    ForbiddenError:
      allOf:
        - $ref: '#/components/schemas/BasicErrorModel'
        - type: object
          properties:
            status:
              example: 403
            reason:
              example: Forbidden
    NotFoundError:
      allOf:
        - $ref: '#/components/schemas/BasicErrorModel'
        - type: object
          properties:
            status:
              example: 404
            reason:
              example: Not Found
...
    InternalServerError:
      allOf:
        - $ref: '#/components/schemas/BasicErrorModel'
        - type: object
          properties:
            status:
              example: 500
            reason:
              example: Internal Server Error